Use long instead of int for counter value (#4)
Change counter value type from int to long across domain model, events,
API response, and tests. Overflow-to-zero now occurs at Long.MAX_VALUE
instead of Integer.MAX_VALUE.

Resolves #3

+++ src/main/java/com/example/domain/CounterEvent.java
@@ -5,10 +5,10 @@ import akka.javasdk.annotations.TypeName;
 public sealed interface CounterEvent {
 
   @TypeName("value-incremented")
-  record ValueIncremented(int newValue) implements CounterEvent {}
+  record ValueIncremented(long newValue) implements CounterEvent {}
 
   @TypeName("value-decremented")
-  record ValueDecremented(int newValue) implements CounterEvent {}
+  record ValueDecremented(long newValue) implements CounterEvent {}
 
   @TypeName("name-changed")
   record NameChanged(String newName) implements CounterEvent {}
